# Avoid reallocations in base64 encoding